finished implementation of delete-profile functionality and the according design
