+++ client/src/App.js
@@ -11,7 +11,8 @@ import Searchbar from './components/Ui/Searchbar';
 import CreateProfile from './Pages/CreateProfile';
 import ScrollToTop from './hooks/useScrollToTop';
 import FriendsCards from './Pages/FriendsCards';
-import sortFilter from './lib/friendsFunctions';
+import sortFilter from './lib/helperFunctions';
+import { checkExistingUser } from './lib/helperFunctions';
 import Banner from '../src/components/Banner';
 import requests from './services/requests';
 import Watchlist from './Pages/Watchlist';
@@ -38,14 +39,6 @@ function App() {
   const [open, setOpen] = useState(false);
   const [query, setQuery] = useState('');
   const [page, setPage] = useState(1);
-
-  function checkExistingUser(userProfile) {
-    if (userProfile === null) {
-      return false;
-    } else if (Object.keys(userProfile).length === 0) {
-      return false;
-    } else return true;
-  }
 
   useEffect(() => {
     async function getFriends() {
@@ -335,12 +328,12 @@ const MovieWrapper = styled.div`
   align-items: center;
   flex-direction: column;
   margin: 0 auto;
-  padding: 10px;
-  width: 100%;
   max-width: 1020px;
   overflow-x: hidden;
   overflow-y: scroll;
+  padding: 10px;
   scrollbar-width: none;
+  width: 100%;
   ::-webkit-scrollbar {
     display: none;
   }
@@ -393,9 +386,9 @@ const BackArrow = styled(UpArrow)`
 
 const GridWrapper = styled.div`
   display: grid;
-  gap: 20px;
   justify-items: center;
   grid-template-columns: repeat(4, 1fr);
+  gap: 20px;
   max-width: 1020px;
   @media (max-width: 800px) {
     grid-template-columns: repeat(3, 1fr);
@@ -411,9 +404,9 @@ const GridWrapper = styled.div`
 const ProfileWrapper = styled.div`
   margin: 0 auto;
   max-width: 1020px;
-  width: 100%;
   overflow-x: hidden;
   scrollbar-width: none;
+  width: 100%;
   ::-webkit-scrollbar {
     display: none;
   }
@@ -422,15 +415,15 @@ const ProfileWrapper = styled.div`
 const FriendsHeadline = styled.h2`
   display: flex;
   margin-left: 20px;
-  width: 100%;
   max-width: 1020px;
+  width: 100%;
 `;
 
 const FriendsFlex = styled.div`
-  margin-left: 25px;
   display: flex;
-  padding: 20px;
+  margin-left: 25px;
   overflow-x: scroll;
+  padding: 20px;
   scrollbar-width: none;
   ::-webkit-scrollbar {
     display: none;
